Break up TabBar.swift
